Added first and last name validation
