+++ index.html
@@ -17,14 +17,7 @@
 <div class="box">
 <div class="head">
   <form id="form">
-    <div class="form-control">
       <h2>Create Account</h2>
-          <!--<label class="firstName">First Name:</label><br>
-            <input type="text" placeholder="Carlos" value="" onfocusout="validateFirstName"><br><!-- user input name -->
-          <!--<label class="lastName">Last Name:</label><br>
-            <input type="text" placeholder="Williams" value="" onfocusout="validateLastName"><br><!-- user input name -->
-<!-- Remove below this -->
-          <!-- Lexie - First Name -->
           <div class="form-control">
             <label class="name">First Name:</label><br>
             <input type="text" placeholder="Jane" value="" id="fname">
@@ -32,7 +25,6 @@
               <i class="fas fa-exclamation-circle"></i>
               <small></small>
           </div>
-            <!-- Lexie - Last Name -->
           <div class="form-control">
             <label class="name">Last Name:</label><br>
             <input type="text" placeholder="Doe" value="" id="lname">
@@ -40,25 +32,19 @@
               <i class="fas fa-exclamation-circle"></i>
               <small></small>
           </div>
-<!-- Remove above this -->
           <div class="form-control">
-            <label class="num">Number:</label><br><!-- user number required for login -->
+            <label class="phone-number">Number:</label><br>
               <input type="text" placeholder="xxx-xxx-xxxx" value="" id="phonenumber">
-              <!-- Success and Error Icons -->
                 <i class="fas fa-check-circle"></i>
                 <i class="fas fa-exclamation-circle"></i>
               <small></small>
-              <!-- Submit Button -->
-              <br><br>
-          <button type="submit" id="submit">Submit</button>
           </div>
-    </div>
-    
+        <br>
+      <button type="submit" id="submit">Submit</button>
   </form>
 </div>       
 </div>
 
-<script src='phone.js'></script>
 <script src='admin.js'></script>
 <script src='index.js'></script>
